feat: support the detection of mono repo tags when explicitly provided (#34)
Closes #30

A mono repo will typically use a semantic version tag structure of a
prefix followed by the version. This allows the tag to associated with a
dedicated component within the mono repo. If a tag is provided in this
format, the approach to scanning must change. Only ancestor tags with
the same prefix should be considered.