feat: always filter commits based on the provided directory within the repository
Closes #35

Generating a release note from a sub-directory within a git project
should not include all commits from the projects history. This can
result in a note containing unassociated commits. When querying the git
log with a sub-directory, e.g. git log ui/, the history is filtered to
only commits that modify files within that part of the tree. By
switching to git2::Repository::discover the current location within a
repository can be determined and the filtering dynamically changed to
mimic this behavior.